+++ include/beast/http/string_body.hpp
@@ -43,6 +43,11 @@ private:
         }
 
         void
+        init(error_code&) noexcept
+        {
+        }
+
+        void
         write(void const* data,
             std::size_t size, error_code&) noexcept
         {